Regenerate the low-level Rust bindings using `witx-bindgen` (#67)
* Re-generate the Rust bindings using `witx-bindgen`

This had not been done in a long time, so `witx-bingen`'s output format
has changed slightly since then.

* Add `scripts/regenerate-bindings-from-witx`

This script simplifies the process of regenerating the low-level Rust
bindings from a WITX file. The defaults chosen in the script are the
latest as of this commit date.

* ci: check that Rust bindings do not need to be regenerated

As the `witx-bindgen` and the `wasi-nn` specification change, we would
like to know if our checked-in code that has been auto-generated needs
to be generated again. To do so, this changes the default revision to
pull for each of these sources to the `main` branch.