Ignore files in their correct directory (#78)
Previously, a top-level `.gitignore` meant that ignored files in the
language sub-directories (`assemblyscript` and `rust`) were still
included in the packaged files. This drastically increased the size of
the packages. By moving the ignored files down a level and adding an
`.npmignore` for the AssemblyScript library, the packages can remain
small.