Update Rust gen to use .wit spec
This commit updates the regeneration script for the Rust bindings to
use `wasi-nn.wit` types instead of `wasi-nn.witx`.

The scripts has also been run and the bindings (src/generated.rs) have
been updated.

This example currently compiles but fails at runtime because the version
of the wasi-nn spec that wasmtime is using is not the latest. Once it
has been updated this example should work (I've tested it with the
version that wasmtime is currently on and it works but that requires
changes to the examples in this repository).

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>